for react query, add unit tests with rendering to test if component is failing on fetching data when doing SSR
