+++ packages/xarc-react-query/test/spec/node-index.spec.tsx
@@ -1,13 +1,11 @@
 /* eslint-disable prefer-arrow-callback */
 import React from "react"; // eslint-disable-line
-//
-import { reactQueryFeature } from "../../src/node/index";
 import { describe, it } from "mocha";
 import { expect } from "chai";
 import { renderToString } from "react-dom/server";
-import { SubAppDef, envHooks } from "@xarc/subapp";
-import { SubAppContainer } from "@xarc/subapp";
+import { SubAppDef, SubAppContainer, envHooks } from "@xarc/subapp";
 import { useQuery } from "react-query";
+import { reactQueryFeature } from "../../src/node/index";
 import { testFetch } from "../prefetch";
 
 const { createElement } = React; // eslint-disable-line
@@ -37,7 +35,7 @@ describe("reactQueryFeature node.js", function () {
     expect(def._features.reactQuery).to.be.an("object");
   });
 
-  it("should render subapp with react-query", async () => {
+  it("should render subapp with react-query if it successfully fetches data when doing SSR", async () => {
     const container = new SubAppContainer({});
 
     envHooks.getContainer = () => container;
@@ -71,6 +69,47 @@ describe("reactQueryFeature node.js", function () {
 
     expect(str).equals(
       `<div>test <!-- -->{&quot;msg&quot;:&quot;foo&quot;,&quot;queryKey&quot;:[&quot;test&quot;]}</div>`
+    );
+  });
+
+  it("should render subapp with react-query if it fails on fetching data when doing SSR", async () => {
+    const container = new SubAppContainer({});
+
+    envHooks.getContainer = () => container;
+
+    const factory = reactQueryFeature({
+      React,
+      serverModule: undefined
+    });
+
+    const def = {
+      name: "test",
+      getModule() {
+        return Promise.resolve({});
+      },
+      _features: {}
+    } as SubAppDef;
+
+    container.declare("test", def);
+
+    factory.add(def);
+
+    (def._features.reactQuery as any).wrap = props => {
+      return JSON.stringify(props);
+    };
+
+    const result = await def._features.reactQuery.execute({
+      input: {
+        Component: () => {
+          const { data } = useQuery("test", testFetch);
+          return <div>test {JSON.stringify(data)}</div>;
+        }
+      }
+    });
+
+    const mockComponentFunc = result.Component;
+    expect(mockComponentFunc()).equals(
+      `{"queryClient":{"queryCache":{"listeners":[],"config":{},"queries":[],"queriesMap":{}},"mutationCache":{"listeners":[],"config":{},"mutations":[],"mutationId":0},"defaultOptions":{},"queryDefaults":[],"mutationDefaults":[]},"dehydratedState":{"mutations":[],"queries":[]}}`
     );
   });
 });
